requirements.txt
streamlit
pillow
python-docx
pdf2image
PyPDF2
pdf2docx
python-pptx
pytesseract
transformers
pandas
pydub
openpyxl
python-magic
fpdf
streamlit>=1.35.0
pdf2docx>=0.5.6
PyPDF2>=3.0.1
python-docx>=1.1.0
pillow>=10.3.0
pydub>=0.25.1
openpyxl>=3.1.2
pandas>=2.2.2
python-pptx>=0.6.21
SpeechRecognition>=3.10.0
openai-whisper
docx2pdf
pdf2docx
PyMuPDF
pytesseract
python-pptx
fpdf
comtypes ; sys_platform == 'win32'



